moved NullValue type to separate file
