+++ src/Postgres.Marula.Calculations/ParameterValues/Base/IParameterValue.cs
@@ -1,7 +1,5 @@
-using System;
 using Postgres.Marula.Calculations.ParameterProperties;
 using Postgres.Marula.Calculations.Parameters.Base;
-using Postgres.Marula.Infrastructure.TypeDecorators;
 
 namespace Postgres.Marula.Calculations.ParameterValues.Base
 {
@@ -19,31 +17,5 @@ namespace Postgres.Marula.Calculations.ParameterValues.Base
 		/// Parameter unit.
 		/// </summary>
 		IUnit Unit { get; }
-	}
-
-	/// <summary>
-	/// Represents absence of value.
-	/// </summary>
-	public sealed class NullValue : IParameterValue
-	{
-		private NullValue(NonEmptyString parameterName)
-			=> ParameterLink = new ParameterLink(parameterName);
-
-		/// <summary>
-		/// Create new instance of <see cref="NullValue"/> related to <paramref name="parameter"/>. 
-		/// </summary>
-		public static IParameterValue OfParameter(IParameter parameter)
-			=> new NullValue(parameter.Name);
-
-		/// <inheritdoc />
-		public IParameterLink ParameterLink { get; }
-
-		/// <inheritdoc />
-		IUnit IParameterValue.Unit => throw AccessError();
-
-		/// <summary>
-		/// Get exception object to throw on any member access. 
-		/// </summary>
-		private static InvalidOperationException AccessError() => throw new($"Unable to access {nameof(NullValue)} members.");
 	}
 }
