+++ src/Postgres.Marula.Calculations/ParametersManagement/PgSettings.cs
@@ -8,6 +8,7 @@ using Postgres.Marula.Calculations.Configuration;
 using Postgres.Marula.Calculations.ExternalDependencies;
 using Postgres.Marula.Calculations.ParameterProperties;
 using Postgres.Marula.Calculations.Parameters.Base;
+using Postgres.Marula.Calculations.ParameterValues;
 using Postgres.Marula.Calculations.ParameterValues.Base;
 using Postgres.Marula.Calculations.ParameterValues.Parsing;
 using Postgres.Marula.Infrastructure.Extensions;
