+++ src/Postgres.Marula.Calculations/Pipeline/MiddlewareComponents/ValueCalculationsMiddleware.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Logging;
 using PipelineNet.Middleware;
 using Postgres.Marula.Calculations.Parameters.Base;
 using Postgres.Marula.Calculations.ParametersManagement;
+using Postgres.Marula.Calculations.ParameterValues;
 using Postgres.Marula.Calculations.ParameterValues.Base;
 using Postgres.Marula.Infrastructure.Extensions;
